+++ extensions/vscode_workspace/src/project-config/validator.ts
@@ -1,0 +1,10 @@
+import type { Static } from "@sinclair/typebox";
+import type { Any } from "ts-toolbelt";
+import { ProjectConfigSchema, TaskSchema } from "./schema";
+import { compileSchema } from "./utils";
+
+export type Task = Any.Compute<Static<typeof TaskSchema>>;
+export const TaskValidator = compileSchema(TaskSchema);
+
+export type ProjectConfig = Any.Compute<Static<typeof ProjectConfigSchema>>;
+export const ProjectConfigValidator = compileSchema(ProjectConfigSchema);