+++ extensions/vscode_workspace/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
-## UNRELEASED
+## 0.0.1
 
 - Release under MIT license
+- Attach terminals to workspace tasks
+- Installs extensions specified in the workspace config